hashmap: rework iterator macros to avoid declaring type in for loop
GCC permits declaring an anonymous struct in a for loop, but it is a
gray area in the C language specification, and Clang and MSVC do
not permit it and fail the compilation.

To fix this issue, I added the iterator type to the hashmap struct
definition. This required modifying the iterator type to avoid directly
referencing the hashmap struct.